+++ tests/src/secure/tests_ubi_secure_fault_injection.c
@@ -69,11 +69,8 @@ static void ztest_testcase_teardown(void *ctx)
 
 static struct ubi_device *sec_init(void)
 {
-	static struct ubi_crypto_config cfg;
-	cfg = ubi_test_mock_crypto_config();
-	struct ubi_device *ubi = NULL;
+	struct ubi_device *const ubi = ubi_test_secure_init(&flash);
 
-	zassert_ok(ubi_device_init(&flash, &cfg, &ubi));
 	g_ubi = ubi;
 	return ubi;
 }